implemented logging in
