+++ src/pages/Login.jsx
@@ -1,9 +1,12 @@
 import { useState } from "react"
 import { getUserByUsername } from "../services/userService"
+import { useNavigate } from "react-router-dom"
 
 const Login = () =>
 {
     const [username, setUsername] = useState("")
+
+    const navigate = useNavigate()
 
     const onUsernameInputChanged = (event) =>
     {
@@ -17,10 +20,8 @@ const Login = () =>
             {
                 if(user.length === 1)
                 {
-                    localStorage.setItem(
-                        "guitar-practicer-user",
-                        JSON.stringify(user[0])
-                    )
+                    window.localStorage.setItem("guitar-practicer-user", JSON.stringify(user[0]))
+                    navigate("/")
                 } else
                 {
 
@@ -33,8 +34,8 @@ const Login = () =>
         <h1 className="text-6xl">Log In</h1>
         <div className="space-x-5">
             <label name="username">username:</label>
-            <input name="username" type="text" onChange={onUsernameInputChanged} />
-            <button>Log In</button>
+            <input name="username" type="text" className="text-black" onChange={onUsernameInputChanged} />
+            <button onClick={onLoginButtonClicked}>Log In</button>
         </div>
         <div className="flex">
             <p>Don't have an account?  </p>
